Fix wrapper
