+++ wrappers/node/src/index.ts
@@ -11,7 +11,8 @@ export * from './api/disclosed-proof';
 export * from './api/utils';
 export * from './api/wallet';
 export * from './api/agent';
-export * from './api/out-of-band';
+export * from './api/out-of-band-sender';
+export * from './api/out-of-band-receiver';
 export * from './api/vcx-payment-txn';
 export * from './vcx';
 export * from './rustlib';
